Gen 565 private twab controller has fairly low limits

+++ src/TwabController.sol
@@ -16,10 +16,10 @@ error PeriodLengthTooShort();
 
 /// @notice Emitted when the period offset is not in the past.
 /// @param periodOffset The period offset that was passed in
-error PeriodOffsetInFuture(uint32 periodOffset);
+error PeriodOffsetInFuture(uint48 periodOffset);
 
 // The minimum period length
-uint32 constant MINIMUM_PERIOD_LENGTH = 1 hours;
+uint48 constant MINIMUM_PERIOD_LENGTH = 1 hours;
 
 address constant SPONSORSHIP_ADDRESS = address(1);
 
@@ -41,11 +41,11 @@ contract TwabController {
   /// @notice Allows users to revoke their chances to win by delegating to the sponsorship address.
   
   /// @notice Sets the minimum period length for Observations. When a period elapses, a new Observation is recorded, otherwise the most recent Observation is updated.
-  uint32 public immutable PERIOD_LENGTH;
+  uint48 public immutable PERIOD_LENGTH;
 
   /// @notice Sets the beginning timestamp for the first period. This allows us to maximize storage as well as line up periods with a chosen timestamp.
   /// @dev Ensure that the PERIOD_OFFSET is in the past.
-  uint32 public immutable PERIOD_OFFSET;
+  uint48 public immutable PERIOD_OFFSET;
 
   /* ============ State ============ */
 
@@ -70,8 +70,8 @@ contract TwabController {
   event IncreasedBalance(
     address indexed vault,
     address indexed user,
-    uint96 amount,
-    uint96 delegateAmount
+    uint112 amount,
+    uint112 delegateAmount
   );
 
   /**
@@ -84,8 +84,8 @@ contract TwabController {
   event DecreasedBalance(
     address indexed vault,
     address indexed user,
-    uint96 amount,
-    uint96 delegateAmount
+    uint112 amount,
+    uint112 delegateAmount
   );
 
   /**
@@ -120,7 +120,7 @@ contract TwabController {
    * @param amount the amount the total supply increased by
    * @param delegateAmount the amount the delegateTotalSupply increased by
    */
-  event IncreasedTotalSupply(address indexed vault, uint96 amount, uint96 delegateAmount);
+  event IncreasedTotalSupply(address indexed vault, uint112 amount, uint112 delegateAmount);
 
   /**
    * @notice Emitted when the total supply or delegateTotalSupply is decreased.
@@ -128,7 +128,7 @@ contract TwabController {
    * @param amount the amount the total supply decreased by
    * @param delegateAmount the amount the delegateTotalSupply decreased by
    */
-  event DecreasedTotalSupply(address indexed vault, uint96 amount, uint96 delegateAmount);
+  event DecreasedTotalSupply(address indexed vault, uint112 amount, uint112 delegateAmount);
 
   /**
    * @notice Emited when a Total Supply Observation is recorded to the Ring Buffer.
@@ -156,7 +156,7 @@ contract TwabController {
    * @param _periodOffset Sets the beginning timestamp for the first period. This allows us to maximize storage as well
    *      as line up periods with a chosen timestamp.
    */
-  constructor(uint32 _periodLength, uint32 _periodOffset) {
+  constructor(uint48 _periodLength, uint48 _periodOffset) {
     if (_periodLength < MINIMUM_PERIOD_LENGTH) {
       revert PeriodLengthTooShort();
     }
@@ -252,7 +252,7 @@ contract TwabController {
   function getBalanceAt(
     address vault,
     address user,
-    uint32 periodEndOnOrAfterTime
+    uint48 periodEndOnOrAfterTime
   ) external view returns (uint256) {
     TwabLib.Account storage _account = userObservations[vault][user];
     return TwabLib.getBalanceAt(PERIOD_LENGTH, PERIOD_OFFSET, _account.observations, _account.details, _periodEndOnOrAfter(periodEndOnOrAfterTime));
@@ -264,7 +264,7 @@ contract TwabController {
    * @param periodEndOnOrAfterTime The time in the past for which the balance is being queried. The time will be snapped to a period end time on or after the timestamp.
    * @return The total supply at the target time
    */
-  function getTotalSupplyAt(address vault, uint32 periodEndOnOrAfterTime) external view returns (uint256) {
+  function getTotalSupplyAt(address vault, uint48 periodEndOnOrAfterTime) external view returns (uint256) {
     TwabLib.Account storage _account = totalSupplyObservations[vault];
     return TwabLib.getBalanceAt(PERIOD_LENGTH, PERIOD_OFFSET, _account.observations, _account.details, _periodEndOnOrAfter(periodEndOnOrAfterTime));
   }
@@ -281,8 +281,8 @@ contract TwabController {
   function getTwabBetween(
     address vault,
     address user,
-    uint32 startTime,
-    uint32 endTime
+    uint48 startTime,
+    uint48 endTime
   ) external view returns (uint256) {
     TwabLib.Account storage _account = userObservations[vault][user];
     // We snap the timestamps to the period end on or after the timestamp because the total supply records will be sparsely populated.
@@ -308,8 +308,8 @@ contract TwabController {
    */
   function getTotalSupplyTwabBetween(
     address vault,
-    uint32 startTime,
-    uint32 endTime
+    uint48 startTime,
+    uint48 endTime
   ) external view returns (uint256) {
     TwabLib.Account storage _account = totalSupplyObservations[vault];
     // We snap the timestamps to the period end on or after the timestamp because the total supply records will be sparsely populated.
@@ -330,7 +330,7 @@ contract TwabController {
    * @param _timestamp The timestamp to check
    * @return The end timestamp of the period that ends on or immediately after the given timestamp
    */
-  function periodEndOnOrAfter(uint32 _timestamp) external view returns (uint32) {
+  function periodEndOnOrAfter(uint48 _timestamp) external view returns (uint48) {
     return _periodEndOnOrAfter(_timestamp);
   }
 
@@ -339,7 +339,7 @@ contract TwabController {
    * @param _timestamp The timestamp to compute the period end time for
    * @return A period end time.
    */
-  function _periodEndOnOrAfter(uint32 _timestamp) internal view returns (uint32) {
+  function _periodEndOnOrAfter(uint48 _timestamp) internal view returns (uint48) {
     if (_timestamp < PERIOD_OFFSET) {
       return PERIOD_OFFSET;
     }
@@ -414,7 +414,7 @@ contract TwabController {
    * @param time The timestamp to check
    * @return period The period the timestamp falls into
    */
-  function getTimestampPeriod(uint32 time) external view returns (uint32) {
+  function getTimestampPeriod(uint48 time) external view returns (uint48) {
     return TwabLib.getTimestampPeriod(PERIOD_LENGTH, PERIOD_OFFSET, time);
   }
 
@@ -423,7 +423,7 @@ contract TwabController {
    * @param time The timestamp to check
    * @return True if the given time is finalized, false if it's during the current overwrite period.
    */
-  function hasFinalized(uint32 time) external view returns (bool) {
+  function hasFinalized(uint48 time) external view returns (bool) {
     return TwabLib.hasFinalized(PERIOD_LENGTH, PERIOD_OFFSET, time);
   }
 
@@ -432,7 +432,7 @@ contract TwabController {
    * @dev The overwrite period is the period during which observations are collated.
    * @return period The timestamp at which the current overwrite period started.
    */
-  function currentOverwritePeriodStartedAt() external view returns (uint32) {
+  function currentOverwritePeriodStartedAt() external view returns (uint48) {
     return TwabLib.currentOverwritePeriodStartedAt(PERIOD_LENGTH, PERIOD_OFFSET);
   }
 
@@ -446,7 +446,7 @@ contract TwabController {
    * @param _to The address to mint balance and delegateBalance to
    * @param _amount The amount to mint
    */
-  function mint(address _to, uint96 _amount) external {
+  function mint(address _to, uint112 _amount) external {
     _transferBalance(msg.sender, address(0), _to, _amount);
   }
 
@@ -458,7 +458,7 @@ contract TwabController {
    * @param _from The address to burn balance and delegateBalance from
    * @param _amount The amount to burn
    */
-  function burn(address _from, uint96 _amount) external {
+  function burn(address _from, uint112 _amount) external {
     _transferBalance(msg.sender, _from, address(0), _amount);
   }
 
@@ -470,7 +470,7 @@ contract TwabController {
    * @param _to The address to transfer balance and delegateBalance to
    * @param _amount The amount to transfer
    */
-  function transfer(address _from, address _to, uint96 _amount) external {
+  function transfer(address _from, address _to, uint112 _amount) external {
     _transferBalance(msg.sender, _from, _to, _amount);
   }
 
@@ -504,7 +504,7 @@ contract TwabController {
    * @param _to the address to which the balance is being transferred
    * @param _amount the amount of balance being transferred
    */
-  function _transferBalance(address _vault, address _from, address _to, uint96 _amount) internal {
+  function _transferBalance(address _vault, address _from, address _to, uint112 _amount) internal {
     if (_to == SPONSORSHIP_ADDRESS) {
       revert CannotTransferToSponsorshipAddress();
     }
@@ -609,7 +609,7 @@ contract TwabController {
     address _vault,
     address _fromDelegate,
     address _toDelegate,
-    uint96 _amount
+    uint112 _amount
   ) internal {
     // If we are transferring tokens from a delegated account to an undelegated account
     if (_fromDelegate != address(0) && _fromDelegate != SPONSORSHIP_ADDRESS) {
@@ -672,8 +672,8 @@ contract TwabController {
   function _increaseBalances(
     address _vault,
     address _user,
-    uint96 _amount,
-    uint96 _delegateAmount
+    uint112 _amount,
+    uint112 _delegateAmount
   ) internal {
     TwabLib.Account storage _account = userObservations[_vault][_user];
 
@@ -711,8 +711,8 @@ contract TwabController {
   function _decreaseBalances(
     address _vault,
     address _user,
-    uint96 _amount,
-    uint96 _delegateAmount
+    uint112 _amount,
+    uint112 _delegateAmount
   ) internal {
     TwabLib.Account storage _account = userObservations[_vault][_user];
 
@@ -756,8 +756,8 @@ contract TwabController {
    */
   function _decreaseTotalSupplyBalances(
     address _vault,
-    uint96 _amount,
-    uint96 _delegateAmount
+    uint112 _amount,
+    uint112 _delegateAmount
   ) internal {
     TwabLib.Account storage _account = totalSupplyObservations[_vault];
 
@@ -800,8 +800,8 @@ contract TwabController {
    */
   function _increaseTotalSupplyBalances(
     address _vault,
-    uint96 _amount,
-    uint96 _delegateAmount
+    uint112 _amount,
+    uint112 _delegateAmount
   ) internal {
     TwabLib.Account storage _account = totalSupplyObservations[_vault];
